requirements.txt
asgiref==3.11.0
Django==5.2.8
django-cors-headers==4.9.0
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
gunicorn==25.0.3
inflection==0.5.1
packaging==26.0
pillow==12.0.0
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
whitenoise==6.11.0
